msg: make 'anon' state explicit
Previously, we were marking an 'anon' Connection (i.e., the Connection used
when a monitor is proxying a Message from another mon) by setting msgr
to NULL.  This could trip up callers of mark_down() or other members that
now try to dereference the msgr pointer.

Instead, make the anon state explicit, so that the magic NULL value is
not as confusing and msgr is not NULL.

This doesn't change that methods like mark_down and send_message are
effectively no-ops on this sort of Connection; that would mean a
significant refactor or redesign.

Signed-off-by: Sage Weil <sage@inktank.com>