osd: calculate a per-pg readable_until map
For a given PG, calculate a readable_until map that, for each interval
that may still be relevant, includes a current lower bound on the amount
of time we will still be allowed to service reads.  This value is based on
the heartbeats we've responded to (as a non-primary) or gotten replies to
(as a primary).  For a stray, we know nothing.

Signed-off-by: Sage Weil <sage@redhat.com>